fixes requested issue removed empty auth file shift flash to app and logo to header
